+++ tests/src/test/spark331/scala/org/apache/spark/sql/RapidsSQLTestsBaseTrait.scala
@@ -32,7 +32,7 @@
 /*** spark-rapids-shim-json-lines
 {"spark": "331"}
 spark-rapids-shim-json-lines ***/
-package org.apache.spark.sql.rapids
+package org.apache.spark.sql
 
 import java.util.{Locale, TimeZone}
