+++ tests/src/test/spark331/scala/org/apache/spark/sql/rapids/suites/RapidsCastSuite.scala
@@ -17,8 +17,9 @@
 /*** spark-rapids-shim-json-lines
 {"spark": "331"}
 spark-rapids-shim-json-lines ***/
-package org.apache.spark.sql
+package org.apache.spark.sql.rapids.suites
 
+import org.apache.spark.sql.RapidsTestsTrait
 import org.apache.spark.sql.catalyst.expressions.{Cast, CastBase, CastSuite, Expression, Literal}
 import org.apache.spark.sql.types._
