+++ tests/src/test/spark331/scala/org/apache/spark/sql/rapids/suites/RapidsDataFrameAggregateSuite.scala
@@ -17,8 +17,17 @@
 /*** spark-rapids-shim-json-lines
 {"spark": "331"}
 spark-rapids-shim-json-lines ***/
-package org.apache.spark.sql
+package org.apache.spark.sql.rapids.suites
 
-import org.apache.spark.sql.catalyst.expressions.StringExpressionsSuite
+import org.apache.spark.sql.{DataFrameAggregateSuite, RapidsSQLTestsTrait}
 
-class RapidsStringExpressionsSuite extends StringExpressionsSuite with RapidsTestsTrait {}
+class RapidsDataFrameAggregateSuite extends DataFrameAggregateSuite with RapidsSQLTestsTrait {
+  // example to show how to replace the logic of an excluded test case in Vanilla Spark
+  testRapids("collect functions" ) {  // "collect functions" was excluded at RapidsTestSettings
+    // println("...")
+  }
+}
+
+
+
+
