+++ tests/src/test/spark331/scala/org/apache/spark/sql/rapids/suites/RapidsJsonFunctionsSuite.scala
@@ -17,7 +17,8 @@
 /*** spark-rapids-shim-json-lines
 {"spark": "331"}
 spark-rapids-shim-json-lines ***/
-package org.apache.spark.sql
+package org.apache.spark.sql.rapids.suites
 
-class RapidsMathFunctionsSuite extends MathFunctionsSuite with RapidsSQLTestsTrait {
-}
+import org.apache.spark.sql.{JsonFunctionsSuite, RapidsSQLTestsTrait}
+
+class RapidsJsonFunctionsSuite extends JsonFunctionsSuite with RapidsSQLTestsTrait {}
