+++ tests/src/test/spark331/scala/org/apache/spark/sql/rapids/suites/RapidsJsonSuite.scala
@@ -17,9 +17,10 @@
 /*** spark-rapids-shim-json-lines
 {"spark": "331"}
 spark-rapids-shim-json-lines ***/
-package org.apache.spark.sql
+package org.apache.spark.sql.rapids.suites
 
 import org.apache.spark.SparkConf
+import org.apache.spark.sql.{sources, RapidsSQLTestsBaseTrait}
 import org.apache.spark.sql.execution.datasources.{InMemoryFileIndex, NoopCache}
 import org.apache.spark.sql.execution.datasources.json.JsonSuite
 import org.apache.spark.sql.execution.datasources.v2.json.JsonScanBuilder
