+++ tests/src/test/spark331/scala/org/apache/spark/sql/rapids/suites/RapidsMathFunctionsSuite.scala
@@ -17,6 +17,9 @@
 /*** spark-rapids-shim-json-lines
 {"spark": "331"}
 spark-rapids-shim-json-lines ***/
-package org.apache.spark.sql
+package org.apache.spark.sql.rapids.suites
 
-class RapidsJsonFunctionsSuite extends JsonFunctionsSuite with RapidsSQLTestsTrait {}
+import org.apache.spark.sql.{MathFunctionsSuite, RapidsSQLTestsTrait}
+
+class RapidsMathFunctionsSuite extends MathFunctionsSuite with RapidsSQLTestsTrait {
+}
